Factories set up + Ballot & AdminUser validations
I'm committing the `spec/examples.txt` file as a way to show that all tests
passed locally after this commit. Maybe this is a good pattern moving forward?

This commit also includes the annotation of all the models in their respective
factory/model/spec files. If we update a model schema in the future we should
remember to run `annotate` from the command line to update the annotations in
these files.